+++ Assets/Plugins/GameJolt/Scripts/API/Constants.cs
@@ -1,11 +1,8 @@
-﻿using UnityEngine;
-using System.Collections;
-
-namespace GameJolt.API
+﻿namespace GameJolt.API
 {
 	public static class Constants
 	{
-		public const string VERSION = "2.2.0";
+		public const string VERSION = "2.3.0";
 
 		public const string SETTINGS_ASSET_NAME = "GJAPISettings";
 		public const string SETTINGS_ASSET_FULL_NAME = SETTINGS_ASSET_NAME + ".asset";